Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo[srv]==4.7.2
motor==3.3.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

